%  file ManlyMix/man/ais.Rd  
%  This file is a component of the package 'ManlyMix' for R
%  copyright (C) 2004-2013 Adelchi Azzalini
%---------------------
\name{acidity}
\alias{acidity}
\docType{data}
\encoding{UTF-8}
\title{Acidity data}
\description{Acidity index measured in a sample of 155 lakes in the Northeastern United States. The data are on the log scale.
}
\usage{data(acidity)}
\format{ 
  A data vector with 155 observations on the acidity index. 
}
\details{The data was first analysed by Crawford et al. (1994).}

\references{
  Crawford, S. L. (1994) \emph{An application of the Laplace method to finite mixture distribution}, Journal of the American Statistical Association, 89, 259-267.
}

\examples{

data(acidity)

}
\keyword{datasets}
